rename: DifficultyLabel 폴더 이동
