+++ src/features/math/components/ProblemTitle/DifficultyLabel/constant.ts
