+++ src/features/math/components/ProblemTitle/DifficultyLabel/index.tsx
